+++ components/ruby/lib/chef_licensing/license_key_fetcher/argument.rb
@@ -16,14 +16,15 @@ module ChefLicensing
         # TODO: this only handles explicit equals
         # TODO: WhyTF are we hand-rolling an option parser
         arg = argv.detect { |a| a.start_with? "--chef-license-key=" }
-        return nil unless arg
+
+        return [] unless arg
 
         match = arg.match(/--chef-license-key=#{LICENSE_KEY_REGEX}/)
         unless match
           raise LicenseKeyNotFetchedError.new("Malformed License Key passed on command line - should be #{LICENSE_KEY_PATTERN_DESC}")
         end
 
-        match[1]
+        [match[1]]
       end
     end
   end